+++ CityController/Mod.cs
@@ -1,5 +1,5 @@
 ﻿using CityController.Systems;
-using CS2Shared.Patch;
+using CS2Shared.Common;
 using CS2Shared.Tools;
 using Game;
 using Game.Modding;
@@ -7,9 +7,9 @@ using System;
 
 namespace CityController;
 
-public class Mod : PatchModBase, IMod {
+public class Mod : ModBase, IMod {
     public override bool BetaVersion => true;
-    public override DateTime VersionDate => new(2024, 9, 9);
+    public override DateTime VersionDate => new(2024, 9, 11);
 
     protected override void CreateSetting() {
         Setting = CityController.Setting.Instance = new Setting(this);