+++ CityController/CHANGELOG.md
@@ -1,4 +1,12 @@
 # Changelog
 
+## [0.1.2] - 2024-09-11
+- Fixed the bug that failed to convert unlimited money to limited money.
+- Updated translation.
+- Added initial money setting.
+
+## [0.1.1] - 2024-09-09
+- Remove Harmony dependency.
+
 ## [0.1.0] - 2024-09-09
 - Initial release.